+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Release date `UNRELEASED`
 
 - Add `:body-params` in non-POST request.
 - (feat api codec breaking) Rename `:param-key-fn` to `:param-name-fn`.
+- (docs) Fix typo in params-request-fn docstring.
 
 ## `1.0.17-beta2`