+++ CHANGELOG.md
@@ -4,13 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.0.18-SNAPSHOT`
+## `1.0.29-SNAPSHOT`
 
 Release date `UNRELEASED`
+
+
+
+## `1.0.28`
+
+Release date `2023-03-26`
 
 - Add `:body-params` in non-POST request.
 - (feat api codec breaking) Rename `:param-key-fn` to `:param-name-fn`.
 - (docs) Fix typo in params-request-fn docstring.
+- (docs bench) Add benchmark data for `util.headers`.
 
 ## `1.0.17-beta2`